requirements.txt
# Core PDF reading
pdfplumber>=0.11.0
PyMuPDF>=1.23.0

# OCR (optional – only needed for scanned PDFs)
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Numerics / data handling
numpy>=1.26.0
pandas>=2.0.0

# Excel output
openpyxl>=3.1.0

# ML-based clustering (optional – used for advanced column detection)
scikit-learn>=1.3.0

# PDF creation for tests (optional)
# reportlab>=4.0.0
